CMake is holding me back...
I need to investigate other build systems. This is stupid. I want to
have the least amount of static content and have the build system
generate the content for me programmatically. Currently this is doable,
but CMake makes it a huge pain in my ass. Every time I think something
is possible, CMake gets in the way. That's not how technology is
supposed to work. Therefore I will probably be ditching CMake soon in
favor of another build system which does not get in my way.