project: properly start new sections on odd pages
This is done by using the 'twoside' modifier on the 'article' document
class and using the 'geometry' package to set the binding offset to 0.
The binding offset reset cleans up the pages for stapled documents which
require no binding offset. This is particularly useful for homework
documents, as those are typically turned in stapled (I'm a student!). I
typically view notes on the computer since it is almost always in a
'draft-like' state, so a null binding offset looks more pleasing in this
medium. When I print the documents I'll set the binding offset to the
default value which probably looks better on a bound medium.